+++ combat_hud.h
@@ -1,7 +1,0 @@
-#ifndef ELORIA_COMBAT_HUD_H
-#define ELORIA_COMBAT_HUD_H
-#include "platform.h"
-void combat_hud_update(const Uint8 *data, int len);
-void display_combat_hud(void);
-void close_combat_hud(void);
-#endif