+++ emotes.h
@@ -1,7 +1,0 @@
-#ifndef __EMOTES_H__
-#define __EMOTES_H__
-
-void send_emote(int emote_id);
-
-#endif
-