+++ marketplace.h
@@ -1,7 +1,0 @@
-#ifndef ELORIA_MARKETPLACE_H
-#define ELORIA_MARKETPLACE_H
-#include "platform.h"
-void marketplace_update(const Uint8 *data, int len);
-void display_marketplace(void);
-void close_marketplace(void);
-#endif